refactor: rename wasm-utils to utils
