+++ .changeset/clever-peaches-do.md
@@ -1,0 +1,12 @@
+---
+"@itk-viewer/remote-viewport": patch
+"@itk-viewer/blosc-zarr": patch
+"@itk-viewer/arcball": patch
+"@itk-viewer/element": patch
+"@itk-viewer/viewer": patch
+"@itk-viewer/utils": patch
+"@itk-viewer/vtkjs": patch
+"@itk-viewer/io": patch
+---
+
+Add orthographic camera and zooming to 2D view.
