+++ packages/utils/src/bounding-box.ts
