+++ packages/viewer/package.json
@@ -46,7 +46,7 @@
   },
   "dependencies": {
     "@itk-viewer/io": "workspace:^",
-    "@itk-viewer/wasm-utils": "workspace:^",
+    "@itk-viewer/utils": "workspace:^",
     "gl-matrix": "^3.4.3",
     "xstate": "5.5.2"
   }
